Added Player/Board/Computer/TicTacToe classes with some functionalities, still unfinished.

+++ src/UI.h
@@ -9,15 +9,15 @@
 //Textures will change after the players click of the buttons
 //Used logic from "Game" class in order to verify the game state
 
-const int TABLE_WIDTH = 3;
-const int TABLE_HEIGHT = 3;
-const int BUTTON_WIDTH = 150;
-const int BUTTON_HEIGHT = 150;
-const int BUTTON_SPACING = 10;
-const int BOARD_PADDING = 100;
-const int BOARD_WIDTH = TABLE_WIDTH * (BUTTON_WIDTH + BUTTON_SPACING) - BUTTON_SPACING + BOARD_PADDING * 2;
-const int BOARD_HEIGHT = TABLE_HEIGHT * (BUTTON_HEIGHT + BUTTON_SPACING) - BUTTON_SPACING + BOARD_PADDING * 2;
-const int LINE_THICKNESS = 10;
+const size_t TABLE_WIDTH = 3;
+const size_t TABLE_HEIGHT = 3;
+const size_t BUTTON_WIDTH = 150;
+const size_t BUTTON_HEIGHT = 150;
+const size_t BUTTON_SPACING = 10;
+const size_t BOARD_PADDING = 100;
+const size_t BOARD_WIDTH = TABLE_WIDTH * (BUTTON_WIDTH + BUTTON_SPACING) - BUTTON_SPACING + BOARD_PADDING * 2;
+const size_t BOARD_HEIGHT = TABLE_HEIGHT * (BUTTON_HEIGHT + BUTTON_SPACING) - BUTTON_SPACING + BOARD_PADDING * 2;
+const size_t LINE_THICKNESS = 10;
 
 class UI {
